Refactor creating separate files for CoreDataFeedImage new classes and mappers.
