+++ FeedStoreChallenge/CoreDataModels/CoreDataFeedImage.swift
@@ -1,0 +1,29 @@
+//
+//  CoreDataFeedImage.swift
+//  FeedStoreChallenge
+//
+//  Created by Guilherme Bueno on 16/05/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import CoreData
+
+@objc(CoreDataFeedImage)
+class CoreDataFeedImage: NSManagedObject {
+	@NSManaged var id: UUID
+	@NSManaged var imageDescription: String?
+	@NSManaged var location: String?
+	@NSManaged var url: URL
+	@NSManaged var feed: CoreDataFeed
+}
+
+extension CoreDataFeedImage {
+	func mapToLocalFeedImage() -> LocalFeedImage {
+		return LocalFeedImage(
+			id: id,
+			description: imageDescription,
+			location: location,
+			url: url
+		)
+	}
+}
